+++ ChangeLog.md
@@ -12,6 +12,7 @@ More detailed information about incremental changes can be found in the
 - Fixed invalidation of cached Hessians when reoptimizing with same structure.
 - Added `OptionsList::SetBoolValue()` and `OptionsList::SetBoolValueIfUnset()`. [#492]
 - Skip check for and link against libdl if linear solver loader is disabled.
+- Fixed missing initialization of `BacktrackingLineSearch::in_watchdog_`.
 
 ### 3.14.0 (2021-06-15)